Code refactoring attempt
I've created :

- data/soccer-page-data.ts to contain all the hardcoded data of the soccer page.

- components/sports/sports-pages-components.tsx to contain all the component that every sport page will share (mostly generic card, etc.)

Then in the components/sports/soccer.tsx I modified the soccer header to use the data in the soccer-page-data.ts file. But somehow it doesn't work (the main issue is the css property). When I inspect the page, the css prpperties are there in the code, but they seems to not apply properly.